refactor: 컨트롤러-서비스 구조가 잘못된것같아서 service 계층을 지움
